fix: close the open findings from the PR review pass
Cap concurrent speaker queues at 32. computeIfAbsent allocated one per
distinct sender_session before anything about the packet was checked, so a
server handing out ids freely outran the ~10 s idle retirement.

Make offer() report rejection instead of swallowing the packet. Retirement
and the removal from the speaker map are not one step, so a packet offered
in between was lost; the reader now takes a fresh queue. The gate is
`retired` rather than close(), because fillTick sets retired first and a
gate on close would still lose everything in between. Removal is by
identity for the same reason — the reader may already have swapped in a
replacement.

Widen the mixer accumulator read to Long. -Int.MIN_VALUE is still
Int.MIN_VALUE, so that one sample kept a negative magnitude, skipped the
limiter, and truncated to 0 — silence at the loudest possible sample.

Export only the four JNI entry points from libdumble.so. Static linking
pulled libopus's whole symbol table in: 233 exports, now 4, and the
stripped arm64 library drops from 192,728 to 154,256 bytes.

Upload a native symbol table with the release bundle now that the .so
files are our own code rather than stripped AndroidX transitives.

+++ app/src/main/java/me/danielstiner/dumble/mumble/voice/AudioMixer.kt
@@ -23,7 +23,11 @@ object AudioMixer {
 
     fun finalizeMix(acc: IntArray, dst: ShortArray, n: Int) {
         for (i in 0 until n) {
-            val x = acc[i]
+            // Widened because -Int.MIN_VALUE is still Int.MIN_VALUE: the magnitude would stay
+            // negative, slip past the threshold test, and truncate to 0 — silence exactly where the
+            // mix is loudest. Overflowing the accumulator takes ~65k concurrent full-scale
+            // speakers, so this is cheap insurance against a branch that cannot be tested for.
+            val x = acc[i].toLong()
             val ax = if (x < 0) -x else x
             dst[i] = if (ax <= THRESHOLD) {
                 x.toShort()
